+++ app/src/main/java/woowacourse/shopping/DIApplication.kt
@@ -2,9 +2,9 @@ package woowacourse.shopping
 
 import android.app.Application
 import com.woowa.di.injection.ModuleRegistry
-import woowacourse.shopping.di.DaoDI
-import woowacourse.shopping.di.DaoModule
-import woowacourse.shopping.di.RepositoryModule
+import woowacourse.shopping.di.dao.DaoDI
+import woowacourse.shopping.di.dao.DaoModule
+import woowacourse.shopping.di.repository.RepositoryModule
 import woowacourse.shopping.di.repository.RepositoryDI
 
 class DIApplication : Application() {
